Use async NBGL C API in Rust SDK (Home&Settings, AddressReview and ReviewStatus)
