+++ ledger_device_sdk/examples/nbgl_address.rs
@@ -23,20 +23,22 @@ extern "C" fn sample_main() {
     }
 
     let mut comm = Comm::new();
+    // Initialize reference to Comm instance for NBGL
+    // API calls.
+    init_comm(&mut comm);
 
     let addr_hex = "0x1234567890ABCDEF1234567890ABCDEF12345678";
 
     // Load glyph from 64x64 4bpp gif file with include_gif macro. Creates an NBGL compatible glyph.
     const FERRIS: NbglGlyph =
         NbglGlyph::from_include(include_gif!("examples/crab_64x64.gif", NBGL));
-
     // Display the address confirmation screen.
-    let success = NbglAddressReview::new(&mut comm)
+    let success = NbglAddressReview::new()
         .glyph(&FERRIS)
         .verify_str("Verify Address")
         .show(addr_hex);
-
-    NbglReviewStatus::new(&mut comm)
+    NbglReviewStatus::new()
         .status_type(StatusType::Address)
         .show(success);
+    exit_app(0);
 }
