+++ ledger_device_sdk/src/nbgl/nbgl_choice.rs
@@ -1,0 +1,58 @@
+use super::*;
+
+/// A wrapper around the synchronous NBGL ux_sync_status C API binding.
+/// Draws a generic choice page, described in a centered info (with configurable icon),
+/// thanks to a button and a footer at the bottom of the page.
+pub struct NbglChoice<'a> {
+    glyph: Option<&'a NbglGlyph<'a>>,
+}
+
+impl<'a> NbglChoice<'a> {
+    pub fn new() -> NbglChoice<'a> {
+        NbglChoice { glyph: None }
+    }
+
+    pub fn glyph(self, glyph: &'a NbglGlyph) -> NbglChoice<'a> {
+        NbglChoice {
+            glyph: Some(glyph),
+            ..self
+        }
+    }
+
+    pub fn show(
+        self,
+        message: &str,
+        sub_message: &str,
+        confirm_text: &str,
+        cancel_text: &str,
+    ) -> bool {
+        unsafe {
+            let icon: nbgl_icon_details_t = match self.glyph {
+                Some(g) => g.into(),
+                None => nbgl_icon_details_t::default(),
+            };
+            let message = CString::new(message).unwrap();
+            let sub_message = CString::new(sub_message).unwrap();
+            let confirm_text = CString::new(confirm_text).unwrap();
+            let cancel_text = CString::new(cancel_text).unwrap();
+
+            let sync_ret = ux_sync_choice(
+                &icon as *const nbgl_icon_details_t,
+                message.as_ptr() as *const c_char,
+                sub_message.as_ptr() as *const c_char,
+                confirm_text.as_ptr() as *const c_char,
+                cancel_text.as_ptr() as *const c_char,
+            );
+
+            // Return true if the user approved the transaction, false otherwise.
+            match sync_ret {
+                UX_SYNC_RET_APPROVED => {
+                    return true;
+                }
+                _ => {
+                    return false;
+                }
+            }
+        }
+    }
+}
