+++ ledger_device_sdk/src/nbgl/nbgl_home_and_settings.rs
@@ -1,0 +1,193 @@
+use super::*;
+
+/// Callback triggered by the NBGL API when a setting switch is toggled.
+unsafe extern "C" fn settings_callback(token: c_int, _index: u8, _page: c_int) {
+    let idx = token - FIRST_USER_TOKEN as i32;
+    if idx < 0 || idx >= SETTINGS_SIZE as i32 {
+        panic!("Invalid token.");
+    }
+
+    let setting_idx: usize = idx as usize;
+
+    match SWITCH_ARRAY[setting_idx].initState {
+        OFF_STATE => SWITCH_ARRAY[setting_idx].initState = ON_STATE,
+        ON_STATE => SWITCH_ARRAY[setting_idx].initState = OFF_STATE,
+        _ => panic!("Invalid state."),
+    }
+
+    if let Some(data) = NVM_REF.as_mut() {
+        let mut switch_values: [u8; SETTINGS_SIZE] = data.get_ref().clone();
+        if switch_values[setting_idx] == OFF_STATE {
+            switch_values[setting_idx] = ON_STATE;
+        } else {
+            switch_values[setting_idx] = OFF_STATE;
+        }
+        data.update(&switch_values);
+    }
+}
+
+/// Informations fields name to display in the dedicated
+/// page of the home screen.
+const INFO_FIELDS: [*const c_char; 2] = [
+    "Version\0".as_ptr() as *const c_char,
+    "Developer\0".as_ptr() as *const c_char,
+];
+
+/// Used to display the home screen of the application, with an optional glyph,
+/// information fields, and settings switches.  
+pub struct NbglHomeAndSettings {
+    app_name: CString,
+    info_contents: Vec<CString>,
+    info_contents_ptr: Vec<*const c_char>,
+    setting_contents: Vec<[CString; 2]>,
+    nb_settings: u8,
+    content: nbgl_content_t,
+    generic_contents: nbgl_genericContents_t,
+    info_list: nbgl_contentInfoList_t,
+    icon: nbgl_icon_details_t,
+}
+
+impl SyncNBGL for NbglHomeAndSettings {}
+
+impl<'a> NbglHomeAndSettings {
+    pub fn new() -> NbglHomeAndSettings {
+        NbglHomeAndSettings {
+            app_name: CString::new("").unwrap(),
+            info_contents: Vec::default(),
+            info_contents_ptr: Vec::default(),
+            setting_contents: Vec::default(),
+            nb_settings: 0,
+            content: nbgl_content_t::default(),
+            generic_contents: nbgl_genericContents_t::default(),
+            info_list: nbgl_contentInfoList_t::default(),
+            icon: nbgl_icon_details_t::default(),
+        }
+    }
+
+    pub fn glyph(self, glyph: &'a NbglGlyph) -> NbglHomeAndSettings {
+        let icon = glyph.into();
+        NbglHomeAndSettings { icon: icon, ..self }
+    }
+
+    pub fn infos(
+        self,
+        app_name: &'a str,
+        version: &'a str,
+        author: &'a str,
+    ) -> NbglHomeAndSettings {
+        let mut v: Vec<CString> = Vec::new();
+        v.push(CString::new(version).unwrap());
+        v.push(CString::new(author).unwrap());
+
+        NbglHomeAndSettings {
+            app_name: CString::new(app_name).unwrap(),
+            info_contents: v,
+            ..self
+        }
+    }
+
+    pub fn settings(
+        self,
+        nvm_data: &'a mut AtomicStorage<[u8; SETTINGS_SIZE]>,
+        settings_strings: &[[&'a str; 2]],
+    ) -> NbglHomeAndSettings {
+        unsafe {
+            NVM_REF = Some(transmute(nvm_data));
+        }
+
+        if settings_strings.len() > SETTINGS_SIZE {
+            panic!("Too many settings.");
+        }
+
+        let v: Vec<[CString; 2]> = settings_strings
+            .iter()
+            .map(|s| [CString::new(s[0]).unwrap(), CString::new(s[1]).unwrap()])
+            .collect();
+
+        NbglHomeAndSettings {
+            nb_settings: settings_strings.len() as u8,
+            setting_contents: v,
+            ..self
+        }
+    }
+
+    pub fn show<T: TryFrom<ApduHeader>>(&mut self) -> Event<T>
+    where
+        Reply: From<<T as TryFrom<ApduHeader>>::Error>,
+    {
+        unsafe {
+            loop {
+                self.info_contents_ptr = self
+                    .info_contents
+                    .iter()
+                    .map(|s| s.as_ptr())
+                    .collect::<Vec<_>>();
+
+                self.info_list = nbgl_contentInfoList_t {
+                    infoTypes: INFO_FIELDS.as_ptr() as *const *const c_char,
+                    infoContents: self.info_contents_ptr[..].as_ptr() as *const *const c_char,
+                    nbInfos: INFO_FIELDS.len() as u8,
+                };
+
+                for (i, setting) in self.setting_contents.iter().enumerate() {
+                    SWITCH_ARRAY[i].text = setting[0].as_ptr();
+                    SWITCH_ARRAY[i].subText = setting[1].as_ptr();
+                    let state = if let Some(data) = NVM_REF.as_mut() {
+                        data.get_ref()[i]
+                    } else {
+                        OFF_STATE
+                    };
+                    SWITCH_ARRAY[i].initState = state;
+                    SWITCH_ARRAY[i].token = (FIRST_USER_TOKEN + i as u32) as u8;
+                    SWITCH_ARRAY[i].tuneId = TuneIndex::TapCasual as u8;
+                }
+
+                self.content = nbgl_content_t {
+                    content: nbgl_content_u {
+                        switchesList: nbgl_pageSwitchesList_s {
+                            switches: &SWITCH_ARRAY as *const nbgl_contentSwitch_t,
+                            nbSwitches: self.nb_settings,
+                        },
+                    },
+                    contentActionCallback: Some(settings_callback),
+                    type_: SWITCHES_LIST,
+                };
+
+                self.generic_contents = nbgl_genericContents_t {
+                    callbackCallNeeded: false,
+                    __bindgen_anon_1: nbgl_genericContents_t__bindgen_ty_1 {
+                        contentsList: &self.content as *const nbgl_content_t,
+                    },
+                    nbContents: if self.nb_settings > 0 { 1 } else { 0 },
+                };
+
+                self.ux_sync_init();
+                nbgl_useCaseHomeAndSettings(
+                    self.app_name.as_ptr() as *const c_char,
+                    &self.icon as *const nbgl_icon_details_t,
+                    core::ptr::null(),
+                    INIT_HOME_PAGE as u8,
+                    &self.generic_contents as *const nbgl_genericContents_t,
+                    &self.info_list as *const nbgl_contentInfoList_t,
+                    core::ptr::null(),
+                    Some(quit_callback),
+                );
+                match self.ux_sync_wait(true) {
+                    SyncNbgl::UxSyncRetApduReceived => {
+                        if let Some(comm) = COMM_REF.as_mut() {
+                            if let Some(value) = comm.check_event() {
+                                return value;
+                            }
+                        }
+                    }
+                    SyncNbgl::UxSyncRetQuitted => {
+                        exit_app(0);
+                    }
+                    _ => {
+                        panic!("Unexpected return value from ux_sync_homeAndSettings");
+                    }
+                }
+            }
+        }
+    }
+}
